+++ flask_admin/contrib/fileadmin/azure.py
@@ -1,15 +1,21 @@
+import io
 import os.path as op
+import time
 from datetime import datetime
 from datetime import timedelta
-from time import sleep
 
 try:
-    from azure.storage.blob import BlobPermissions
-    from azure.storage.blob import BlockBlobService
-except ImportError:
-    BlobPermissions = BlockBlobService = None
+    from azure.core.exceptions import ResourceExistsError
+    from azure.storage.blob import BlobProperties
+    from azure.storage.blob import BlobServiceClient
+except ImportError as e:
+    raise Exception(
+        "Could not import `azure.storage.blob`. "
+        "Enable `azure-blob-storage` integration "
+        "by installing `flask-admin[azure-blob-storage]`"
+    ) from e
 
-from flask import redirect
+import flask
 
 from . import BaseFileAdmin
 
@@ -37,38 +43,30 @@ class AzureStorage:
     _send_file_validity = timedelta(hours=1)
     separator = "/"
 
-    def __init__(self, container_name, connection_string):
+    def __init__(self, blob_service_client: BlobServiceClient, container_name: str):
         """
         Constructor
 
+        :param blob_service_client:
+            BlobServiceClient for the Azure Blob Storage account
+
         :param container_name:
             Name of the container that the files are on.
-
-        :param connection_string:
-            Azure Blob Storage Connection String
         """
-
-        if not BlockBlobService:
-            raise ValueError(
-                "Could not import `azure.storage.blob`. "
-                "Enable `azure-blob-storage` integration "
-                "by installing `flask-admin[azure-blob-storage]`"
-            )
-
+        self._client = blob_service_client
         self._container_name = container_name
-        self._connection_string = connection_string
-        self.__client = None
+        try:
+            self._client.create_container(self._container_name)
+        except ResourceExistsError:
+            pass
 
     @property
-    def _client(self):
-        if not self.__client:
-            self.__client = BlockBlobService(connection_string=self._connection_string)
-            self.__client.create_container(self._container_name, fail_on_exist=False)
-        return self.__client
+    def _container_client(self):
+        return self._client.get_container_client(self._container_name)
 
     @classmethod
-    def _get_blob_last_modified(cls, blob):
-        last_modified = blob.properties.last_modified
+    def _get_blob_last_modified(cls, blob: BlobProperties):
+        last_modified = blob.last_modified
         tzinfo = last_modified.tzinfo
         epoch = last_modified - datetime(1970, 1, 1, tzinfo=tzinfo)
         return epoch.total_seconds()
@@ -90,10 +88,13 @@ class AzureStorage:
 
         path_parts = path.split(self.separator) if path else []
         num_path_parts = len(path_parts)
+
         folders = set()
         files = []
 
-        for blob in self._client.list_blobs(self._container_name, path):
+        container_client = self._client.get_container_client(self._container_name)
+
+        for blob in container_client.list_blobs(path):
             blob_path_parts = blob.name.split(self.separator)
             name = blob_path_parts.pop()
 
@@ -103,13 +104,13 @@ class AzureStorage:
             if blob_is_file_at_current_level and not blob_is_directory_file:
                 rel_path = blob.name
                 is_dir = False
-                size = blob.properties.content_length
+                size = blob.size
                 last_modified = self._get_blob_last_modified(blob)
                 files.append((name, rel_path, is_dir, size, last_modified))
             else:
                 next_level_folder = blob_path_parts[: num_path_parts + 1]
-                folder_name = self.separator.join(next_level_folder)
-                folders.add(folder_name)
+                folder = self.separator.join(next_level_folder)
+                folders.add(folder)
 
         folders.discard(directory)
         for folder in folders:
@@ -125,18 +126,10 @@ class AzureStorage:
     def is_dir(self, path):
         path = self._ensure_blob_path(path)
 
-        num_blobs = 0
-        for blob in self._client.list_blobs(self._container_name, path):
-            blob_path_parts = blob.name.split(self.separator)
-            is_explicit_directory = blob_path_parts[-1] == self._fakedir
-            if is_explicit_directory:
+        blobs = self._container_client.list_blobs(name_starts_with=path)
+        for blob in blobs:
+            if blob.name != path:
                 return True
-
-            num_blobs += 1
-            path_cannot_be_leaf = num_blobs >= 2
-            if path_cannot_be_leaf:
-                return True
-
         return False
 
     def path_exists(self, path):
@@ -145,12 +138,13 @@ class AzureStorage:
         if path == self.get_base_path():
             return True
 
-        try:
-            next(iter(self._client.list_blobs(self._container_name, path)))
-        except StopIteration:
+        if path is None:
             return False
-        else:
+
+        # Return true if it exists as either a directory or a file
+        for _ in self._container_client.list_blobs(name_starts_with=path):
             return True
+        return False
 
     def get_base_path(self):
         return ""
@@ -160,80 +154,94 @@ class AzureStorage:
 
         accumulator = []
         breadcrumbs = []
-        for folder in path.split(self.separator):
-            accumulator.append(folder)
-            breadcrumbs.append((folder, self.separator.join(accumulator)))
+        if path is not None:
+            for folder in path.split(self.separator):
+                accumulator.append(folder)
+                breadcrumbs.append((folder, self.separator.join(accumulator)))
         return breadcrumbs
 
     def send_file(self, file_path):
-        file_path = self._ensure_blob_path(file_path)
-
-        if not self._client.exists(self._container_name, file_path):
-            raise ValueError()
-
-        now = datetime.utcnow()
-        url = self._client.make_blob_url(self._container_name, file_path)
-        sas = self._client.generate_blob_shared_access_signature(
-            self._container_name,
-            file_path,
-            BlobPermissions.READ,
-            expiry=now + self._send_file_validity,
-            start=now - self._send_file_lookback,
+        path = self._ensure_blob_path(file_path)
+        if path is None:
+            raise ValueError("No path provided")
+        blob = self._container_client.get_blob_client(path).download_blob()
+        if not blob.properties or not blob.properties.has_key("content_settings"):
+            raise ValueError("Blob has no properties")
+        mime_type = blob.properties["content_settings"]["content_type"]
+        blob_file = io.BytesIO()
+        blob.readinto(blob_file)
+        blob_file.seek(0)
+        return flask.send_file(
+            blob_file, mimetype=mime_type, as_attachment=True, download_name=path
         )
-        return redirect(f"{url}?{sas}")
 
     def read_file(self, path):
         path = self._ensure_blob_path(path)
-
-        blob = self._client.get_blob_to_bytes(self._container_name, path)
-        return blob.content
+        if path is None:
+            raise ValueError("No path provided")
+        blob = self._container_client.get_blob_client(path).download_blob()
+        return blob.readall()
 
     def write_file(self, path, content):
         path = self._ensure_blob_path(path)
-
-        self._client.create_blob_from_text(self._container_name, path, content)
+        if path is None:
+            raise ValueError("No path provided")
+        self._container_client.upload_blob(path, content, overwrite=True)
 
     def save_file(self, path, file_data):
         path = self._ensure_blob_path(path)
-
-        self._client.create_blob_from_stream(
-            self._container_name, path, file_data.stream
-        )
+        if path is None:
+            raise ValueError("No path provided")
+        self._container_client.upload_blob(path, file_data.stream)
 
     def delete_tree(self, directory):
         directory = self._ensure_blob_path(directory)
 
-        for blob in self._client.list_blobs(self._container_name, directory):
-            self._client.delete_blob(self._container_name, blob.name)
+        for blob in self._container_client.list_blobs(directory):
+            self._container_client.delete_blob(blob.name)
 
     def delete_file(self, file_path):
         file_path = self._ensure_blob_path(file_path)
-
-        self._client.delete_blob(self._container_name, file_path)
+        if file_path is None:
+            raise ValueError("No path provided")
+        self._container_client.delete_blob(file_path)
 
     def make_dir(self, path, directory):
         path = self._ensure_blob_path(path)
         directory = self._ensure_blob_path(directory)
-
+        if path is None or directory is None:
+            raise ValueError("No path provided")
         blob = self.separator.join([path, directory, self._fakedir])
         blob = blob.lstrip(self.separator)
-        self._client.create_blob_from_text(self._container_name, blob, "")
+        self._container_client.upload_blob(blob, b"")
 
     def _copy_blob(self, src, dst):
-        src_url = self._client.make_blob_url(self._container_name, src)
-        copy = self._client.copy_blob(self._container_name, dst, src_url)
-        while copy.status != "success":
-            sleep(self._copy_poll_interval_seconds)
-            copy = self._client.get_blob_properties(
-                self._container_name, dst
-            ).properties.copy
+        src_blob_client = self._container_client.get_blob_client(src)
+        dst_blob_client = self._container_client.get_blob_client(dst)
+        copy_result = dst_blob_client.start_copy_from_url(src_blob_client.url)
+        if copy_result.get("copy_status") == "success":
+            return
+
+        for _ in range(10):
+            props = dst_blob_client.get_blob_properties()
+            status = props.copy.status
+            if status == "success":
+                return
+            time.sleep(1)
+
+        if status != "success":
+            props = dst_blob_client.get_blob_properties()
+            copy_id = props.copy.id
+            if copy_id is not None:
+                dst_blob_client.abort_copy(copy_id)
+            raise Exception(f"Copy operation failed: {status}")
 
     def _rename_file(self, src, dst):
         self._copy_blob(src, dst)
         self.delete_file(src)
 
     def _rename_directory(self, src, dst):
-        for blob in self._client.list_blobs(self._container_name, src):
+        for blob in self._container_client.list_blobs(src):
             self._rename_file(blob.name, blob.name.replace(src, dst, 1))
 
     def rename_path(self, src, dst):
@@ -257,15 +265,21 @@ class AzureFileAdmin(BaseFileAdmin):
             Azure Blob Storage Connection String
 
     Sample usage::
-
+        from azure.storage.blob import BlobServiceClient
         from flask_admin import Admin
         from flask_admin.contrib.fileadmin.azure import AzureFileAdmin
 
         admin = Admin()
-
-        admin.add_view(AzureFileAdmin('files_container', 'my-connection-string')
+        client = BlobServiceClient.from_connection_string("my-connection-string")
+        admin.add_view(AzureFileAdmin(client, 'files_container')
     """
 
-    def __init__(self, container_name, connection_string, *args, **kwargs):
-        storage = AzureStorage(container_name, connection_string)
+    def __init__(
+        self,
+        blob_service_client,
+        container_name,
+        *args,
+        **kwargs,
+    ):
+        storage = AzureStorage(blob_service_client, container_name)
         super().__init__(*args, storage=storage, **kwargs)