+++ flask_admin/tests/fileadmin/test_fileadmin_azure.py
@@ -1,6 +1,4 @@
-import os.path as op
-from os import getenv
-from unittest import SkipTest
+import os
 from uuid import uuid4
 
 import pytest
@@ -11,30 +9,28 @@ from .test_fileadmin import Base
 
 
 class TestAzureFileAdmin(Base.FileAdminTests):
-    _test_storage = getenv("AZURE_STORAGE_CONNECTION_STRING")
-
     @pytest.fixture(autouse=True)
     def setup_and_teardown(self):
-        if not azure.BlockBlobService:
-            raise SkipTest("AzureFileAdmin dependencies not installed")
-
+        TEST_STORAGE = os.getenv("AZURE_STORAGE_CONNECTION_STRING")
         self._container_name = f"fileadmin-tests-{uuid4()}"
 
-        if not self._test_storage or not self._container_name:
-            raise SkipTest("AzureFileAdmin test credentials not set")
+        if not TEST_STORAGE or not self._container_name:
+            raise ValueError("AzureFileAdmin test credentials not set, tests will fail")
 
-        client = azure.BlockBlobService(connection_string=self._test_storage)
-        client.create_container(self._container_name)
-        dummy = op.join(self._test_files_root, "dummy.txt")
-        client.create_blob_from_path(self._container_name, "dummy.txt", dummy)
+        self._client = azure.BlobServiceClient.from_connection_string(TEST_STORAGE)
+        self._client.create_container(self._container_name)
+        file_name = "dummy.txt"
+        file_path = os.path.join(self._test_files_root, file_name)
+        blob_client = self._client.get_blob_client(self._container_name, file_name)
+        with open(file_path, "rb") as file:
+            blob_client.upload_blob(file)
 
         yield
 
-        client = azure.BlockBlobService(connection_string=self._test_storage)
-        client.delete_container(self._container_name)
+        self._client.delete_container(self._container_name)
 
     def fileadmin_class(self):
         return azure.AzureFileAdmin
 
     def fileadmin_args(self):
-        return (self._container_name, self._test_storage), {}
+        return (self._client, self._container_name), {}